Update 1-binary_tree_insert_left.c
The need for this function is to insert a new node as the left child of a given parent node. This function takes two arguments: a pointer to the parent node`, and the value to be stored in the new node. The function returns a pointer to the newly created node, or NULL if the operation fails.

This function is meant to first checks if the parent node is NULL, in which case it cannot insert a new node and returns NULL. If the parent node is not NULL, it creates a new node using the binary_tree_node function, passing in the parent node and the given value as arguments. If the creation of the new node fails, the function returns NULL.

If the new node was created successfully, the function checks if the parent already has a left child. If it does, the new node is inserted between the parent and its existing left child. The new node's left pointer is set to the existing left child, and the existing left child's parent pointer is updated to point to the new node. If the parent does not have a left child, the new node is simply set as the parent's left child.

This function can finally return a pointer to the newly created node...