Update 2-binary_tree_insert_right.c
The function first checks whether the parent parameter is NULL, and returns NULL if it is. It then creates a new binary tree node using the binary_tree_node function, passing NULL as the parent parameter to create a node with no parent. If the binary_tree_node function returns NULL, the function also returns NULL.

If the parent node already has a right child, the function updates the parent and right pointers of the new node and the existing right child node, respectively, to insert the new node as the right child of the parent. Otherwise, the function simply sets the right pointer of the parent to the new node and sets the parent pointer of the new node to the parent.

Finally, the function returns a pointer to the new node.